requirements/common.txt
numpy
sortedcontainers
sphinx-bootstrap-theme
# pommerman needs to be installed in a venv
tqdm
pytorch_geometric==1.5.0

git+https://github.com/deepmind/pysc2.git@v3.0#egg=pysc2
git+https://github.com/oxwhirl/smac.git
git+https://github.com/douglasrizzo/python-microRTS.git
